Fixed compilation for GCC 16

+++ src/main/plug/mb_gate.cpp
@@ -908,7 +908,7 @@ namespace lsp
                     bool mute       = (b->pMute->value() >= 0.5f);
                     bool solo       = (enabled) && (b->pSolo->value() >= 0.5f);
                     plug::IPort *sc = (bStereoSplit) ? b->pScSpSource : b->pScSource;
-                    size_t sc_src   = (sc != NULL) ? sc->value() : dspu::SCS_MIDDLE;
+                    size_t sc_src   = (sc != NULL) ? size_t(sc->value()) : size_t(dspu::SCS_MIDDLE);
 
                     b->nScType      = decode_sidechain_type(b->pScType->value());
                     if (b->nScType == SCT_LINK)
